+++ src/network/AhoyNetwork.h
@@ -28,7 +28,6 @@ class AhoyNetwork {
 
             if('\0' == mConfig->sys.deviceName[0])
                 snprintf(mConfig->sys.deviceName, DEVNAME_LEN, "%s", DEF_DEVICE_NAME);
-            WiFi.hostname(mConfig->sys.deviceName);
 
             mAp.setup(&mConfig->sys);
 
@@ -117,13 +116,17 @@ class AhoyNetwork {
 
         void scan(void) {
             mScanActive = true;
-            if(NetworkState::GOT_IP != mStatus)
+            if(mWifiConnecting) {
+                mWifiConnecting = false;
                 WiFi.disconnect();
+            }
             WiFi.scanNetworks(true, true);
         }
         #endif
 
     protected:
+        virtual void setStaticIp() = 0;
+
         void setupIp(std::function<bool(IPAddress ip, IPAddress gateway, IPAddress mask, IPAddress dns1, IPAddress dns2)> cb) {
             if(mConfig->sys.ip.ip[0] != 0) {
                 IPAddress ip(mConfig->sys.ip.ip);
@@ -136,7 +139,7 @@ class AhoyNetwork {
             }
         }
 
-        void OnEvent(WiFiEvent_t event) {
+        virtual void OnEvent(WiFiEvent_t event) {
             switch(event) {
                 case SYSTEM_EVENT_STA_CONNECTED:
                     [[fallthrough]];
@@ -231,6 +234,9 @@ class AhoyNetwork {
         uint32_t *mUtcTimestamp = nullptr;
         bool mConnected = false;
         bool mScanActive = false;
+        #if !defined(ETHERNET)
+        bool mWifiConnecting = false;
+        #endif
 
         OnNetworkCB mOnNetworkCB;
         OnTimeCB mOnTimeCB;