+++ src/utils/spiPatcher.h
@@ -22,11 +22,19 @@ class SpiPatcher {
             // Use binary semaphore instead of mutex for performance reasons
             mutex = xSemaphoreCreateBinaryStatic(&mutex_buffer);
             xSemaphoreGive(mutex);
+            mDev = dev;
+            mBusState = ESP_FAIL;
+        }
 
-            spi_bus_config_t buscfg = {
-                .mosi_io_num = -1,
-                .miso_io_num = -1,
-                .sclk_io_num = -1,
+    public:
+        SpiPatcher(const SpiPatcher &other) = delete;
+        void operator=(const SpiPatcher &) = delete;
+
+        esp_err_t initBus(int mosi = -1, int miso = -1, int sclk = -1, spi_common_dma_t dmaType = SPI_DMA_DISABLED) {
+            mBusConfig = spi_bus_config_t {
+                .mosi_io_num = mosi,
+                .miso_io_num = miso,
+                .sclk_io_num = sclk,
                 .quadwp_io_num = -1,
                 .quadhd_io_num = -1,
                 .data4_io_num = -1,
@@ -37,21 +45,25 @@ class SpiPatcher {
                 .flags = 0,
                 .intr_flags = 0
             };
-            ESP_ERROR_CHECK(spi_bus_initialize(dev, &buscfg, SPI_DMA_DISABLED));
+            ESP_ERROR_CHECK((mBusState = spi_bus_initialize(mDev, &mBusConfig, dmaType)));
+
+            return mBusState;
         }
 
-    public:
-        SpiPatcher(const SpiPatcher &other) = delete;
-        void operator=(const SpiPatcher &) = delete;
-
-        static SpiPatcher* getInstance(spi_host_device_t dev) {
+        static SpiPatcher* getInstance(spi_host_device_t dev, bool initialize = true) {
             if(SPI2_HOST == dev) {
-                if(nullptr == InstanceHost2)
+                if(nullptr == InstanceHost2) {
                     InstanceHost2 = new SpiPatcher(dev);
+                    if(initialize)
+                        InstanceHost2->initBus();
+                }
                 return InstanceHost2;
             } else { // SPI3_HOST
-                if(nullptr == InstanceHost3)
+                if(nullptr == InstanceHost3) {
                     InstanceHost3 = new SpiPatcher(dev);
+                    if(initialize)
+                        InstanceHost3->initBus();
+                }
                 return InstanceHost3;
             }
         }
@@ -59,6 +71,7 @@ class SpiPatcher {
         ~SpiPatcher() { vSemaphoreDelete(mutex); }
 
         inline void addDevice(spi_host_device_t host_id, const spi_device_interface_config_t *dev_config, spi_device_handle_t *handle) {
+            assert(mBusState == ESP_OK);
             if(SPI2_HOST == host_id)
                 mHost2Cnt++;
             if(SPI3_HOST == host_id)
@@ -71,6 +84,7 @@ class SpiPatcher {
         }
 
         inline void request(SpiPatcherHandle* handle) {
+            assert(mBusState == ESP_OK);
             xSemaphoreTake(mutex, portMAX_DELAY);
 
             if (mCurHandle != handle) {
@@ -85,6 +99,7 @@ class SpiPatcher {
         }
 
         inline void release() {
+            assert(mBusState == ESP_OK);
             xSemaphoreGive(mutex);
         }
 
@@ -97,6 +112,9 @@ class SpiPatcher {
         SemaphoreHandle_t mutex;
         StaticSemaphore_t mutex_buffer;
         uint8_t mHost2Cnt = 0, mHost3Cnt = 0;
+        spi_host_device_t mDev = SPI3_HOST;
+        esp_err_t mBusState = ESP_FAIL;
+        spi_bus_config_t mBusConfig;
 };
 
 #endif /*ESP32*/